+++ CHANGELOG.md
@@ -5,4 +5,5 @@
 - first commit
 
 ## February 28, 2023
-- add create ticket command
+
+- add create ticket command (Might need to re-login to Linear in VSCode)